# Speedup string slice decoding.