```
benchmark                     old ns/op     new ns/op     delta
BenchmarkStringSlice-4        323           323           +0.00%
BenchmarkStringSlicePtr-4     901           706           -21.64%
```